Tools and tuning for more accurate state-feedback optimal trajectory tracking.
The main thing we're dealing with in this commit is the observation that
using MPC with the full constraints for large numbers of objects leads
to a large solve time, so we cannot rely on MPC to do fast feedback as
well. Instead, we need to do linear feedback control on the current
state.

After some initial testing it seems like the best option is to use to
built-in optimal feedback controller that can be computed by OCS2's SQP
solver. I've added a lot of tools for visualizing the optimal trajectory
as well as the estimate and commanded trajectories, so I can really know
what's going on.